Create 'update-presets-history' service, clean up old CLI code
