--- cli/utils/get-br-municipalities.js
@@ -1,6 +1,0 @@
-import { municipalitiesCsvFile } from "../../config/paths.js";
-import loadCsv from "../../utils/load-csv.js";
-
-export default async function getBrMunicipalities() {
-  return await loadCsv(municipalitiesCsvFile);
-}
