+++ runner/index.js
@@ -2,8 +2,9 @@ import "dotenv/config";
 import fs from "fs-extra";
 import { program } from "commander";
 import logger from "./helpers/logger.js";
-import { fetchFullHistory } from "../runner/services/fetch-full-history.js";
-import * as citiesOfBrazil from "../runner/services/cities-of-brazil/index.js";
+import { fetchFullHistory } from "./services/fetch-full-history.js";
+import * as citiesOfBrazil from "./services/cities-of-brazil/index.js";
+import { updatePresetsHistory } from "./services/update-presets-history.js";
 
 const pkg = await fs.readJson("./package.json");
 
@@ -19,18 +20,26 @@ program
   .description(
     "Download latest history file and filter it by Osmium tag filters"
   )
-  .action(() => fetchFullHistory(program));
+  .action(fetchFullHistory);
+
+program
+  .command("update-presets-history")
+  .description(
+    "Apply daily diffs to presets history file and update it to present day"
+  )
+  .option("-r, --recursive", "Repeat updates to present day", false)
+  .action(updatePresetsHistory);
 
 program
   .command("setup-cities-of-brazil")
   .description("Setup cities of Brazil workflow")
-  .action(() => citiesOfBrazil.setup(program));
+  .action(citiesOfBrazil.setup);
 
 program
   .command("update-cities-of-brazil")
   .description("Update cities of Brazil")
   .option("-r, --recursive", "Repeat updates to present day", false)
-  .action((options) => citiesOfBrazil.update(options));
+  .action(citiesOfBrazil.update);
 
 program.parse();
